loads items that are not in collection MDM-6980
it happens that the preselection has only a uuid but no further information. So all items that are selected and do not have meta information to display like name are fetched by one collection filter request. The selected models will be updated with the response. Models that are not returned in the response are set to deleted