executes callbacks sequentially instead of parallel
you can register multiple callbacks for an endpoint. Each of them can return a promise that can be resolved or rejected with its own value. The value and the information if it was rejected is passed from one callback to the following registered callback. The last callback that returns a promise defines if the whole promise should be rejected or resolved