fixes wrong implementation
before the callback promises where executed sequentially we used $q to execute all registered callbacks in parallel. $q always returns an array that will passed into the callback. As a quickfix we used the first item of that array but this does not work when multiple promise callbacks are registered on one route. The logic was changed so all registered callbacks of a route will be executed sequentially, therefore only one return value exists and the quickfix can be removed