extends mwWizard functionality
adds directives mwWizardProgress and mwWizardNavigation to make it possible to control the wizard. You don't have to pass a wizard instance to the directive mwWizard anymore, you can use the directive mwWizardNavigation instead. Of course you can still do that if you want to control the wizard from another controller