changes implementation of mwIcon
it is now possible to define icons via the new service mwIcon. The mwIcon directive will use this service to get the icon classes. I paves the way for different icon types like svg, image, font-icon, etc. It makes it also possible to overwrite icons defined in the uikit with custom icons